Remove excess helper class.
